LUN-6: create styled-components theming and refactor styled-component instances
