+++ src/@types/styled.d.ts
@@ -1,0 +1,6 @@
+import 'styled-components'
+import { ThemeType } from '../constants/theme'
+
+declare module 'styled-components' {
+  export interface DefaultTheme extends ThemeType {}
+}
