+++ src/App.tsx
@@ -4,6 +4,8 @@ import Navigator from './navigation/Navigator'
 import React, { ReactElement, useEffect } from 'react'
 import SplashScreen from 'react-native-splash-screen'
 import { SafeAreaProvider } from 'react-native-safe-area-context'
+import { ThemeProvider } from 'styled-components/native'
+import theme from './constants/theme'
 
 const App = (): ReactElement => {
   useEffect(() => {
@@ -13,10 +15,12 @@ const App = (): ReactElement => {
   }, [])
 
   return (
-    <SafeAreaProvider>
-      <StatusBar barStyle='light-content' />
-      <Navigator />
-    </SafeAreaProvider>
+    <ThemeProvider theme={theme}>
+      <SafeAreaProvider>
+        <StatusBar barStyle='light-content' />
+        <Navigator />
+      </SafeAreaProvider>
+    </ThemeProvider>
   )
 }
