+++ src/constants/theme/fonts.ts
@@ -1,0 +1,5 @@
+
+export const FONTS = {
+  contentFontBold: 'SourceSansPro-SemiBold',
+  contentFontRegular: 'SourceSansPro-Regular'
+}
