+++ src/constants/theme/index.ts
@@ -1,0 +1,14 @@
+import { COLORS } from './colors'
+import { FONTS } from './fonts'
+
+export interface ThemeType {
+  colors: typeof COLORS
+  fonts: typeof FONTS
+}
+
+const theme: ThemeType = {
+  colors: COLORS,
+  fonts: FONTS
+}
+
+export default theme
